pyproject.toml
[project]
name = "lineaihelper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "final898y", email = "final898y@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "google-genai>=1.60.0",
    "jinja2>=3.1.6",
    "line-bot-sdk>=3.22.0",
    "loguru>=0.7.3",
    "mypy>=1.19.1",
    "pandas-ta>=0.4.71b0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "yfinance>=1.1.0",
]

[project.scripts]
dev = "lineaihelper.main:start"
test = "lineaihelper.cli:test"
lint = "lineaihelper.cli:lint"
format = "lineaihelper.cli:format"
type-check = "lineaihelper.cli:type_check"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"


[dependency-groups]
dev = [
    "pandas-stubs>=3.0.0.260204",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "types-pyyaml>=6.0.12.20250915",
]
